Updated Level Editor
Pressing numbers Zero -> Four will place different GameObjects on the
map in the level editor.

Pressing the space bar will place a marker on the map(looks like " * ").
If there is a marker on the map and an object is placed, the object
fills every spot on the map between the marker and the cursor. This will
allow you to place many objects of the same type on the map at once.